Validation on Version Number
I'm introducing these changes back as they where as to establish the forking and usage of the circle CI. 
This is in conjunction w/ Pluralsight Tutorial Supercharging Your GitHub Work Flow. 

It should be notes that this user decided not to apply the circle CI to the project outlined in the tutorial as this user did not want the Ruby language to be a barer to me as the user learns.